Some methods.

+++ core/src/main/java/com/google/code/webex/service/impl/MeetingServiceImpl.java
@@ -1,6 +1,21 @@
+/*
+ * Copyright 2010-2011 Nabeel Mukhtar 
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); 
+ * you may not use this file except in compliance with the License. 
+ * You may obtain a copy of the License at 
+ * 
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
+ * See the License for the specific language governing permissions and
+ * limitations under the License. 
+ * 
+ */
 package com.google.code.webex.service.impl;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -33,7 +48,7 @@ public class MeetingServiceImpl extends WebExJaxbService implements
 		LstsummaryMeeting summaryMeeting = factory.createLstsummaryMeeting();
 		LstControlType listControl = new com.webex.schemas._2002._06.service.ObjectFactory().createLstControlType();
 		summaryMeeting.setListControl(listControl);
-		listControl.setMaximumNum(BigInteger.valueOf(5));
+		listControl.setMaximumNum(5L);
 		OrderType order = factory.createOrderType();
 		summaryMeeting.setOrder(order);
 		order.getOrderBy().add(OrderByType.STARTTIME);
